Fetch Remote Data before preparing any plans
